% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGSTimes.R
\name{AGSTimes}
\alias{AGSTimes}
\title{The Analgesic Datasets}
\format{
A data frame with the relief times (in minutes) of 20 patients receiving analgesic.
}
\source{
Jamal, F., Shafiq, S., Aslam, M., Khan, S., Hussain, Z., Abbas, Q. (2024).
Modeling COVID-19 data with a novel neutrosophic Burr-III distribution,
\emph{Scientific Reports}, 14:10810.
}
\description{
Dataset containing times to relief for 20 patients receiving analgesic.
}
\examples{
data("AGSTimes")
AGSTimes
}
