% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloy.R
\name{alloy}
\alias{alloy}
\title{Alloy melting data}
\format{
A data.frame with 18 observations of alloy melting points.
}
\source{
Rao, G. S.  (2003). Neutrosophic Log-Logistic Distribution Model in Complex
Alloy Metal Melting Point Applications, \emph{International Journal of
Computational Intelligence Systems}, 16: 48.

Hassan, M. K. and Aslam, M. (2024). Birnbaum Saunders distribution for
imprecise data: statistical properties, estimation methods, and real life
applications, \emph{scientific reports}, 14: 6955.
}
\description{
It is related to alloy melting points.
}
\examples{
data("alloy")
alloy
}
