% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batteries.R
\name{batteries}
\alias{batteries}
\title{Lifetime of batteries data}
\format{
A data.frame with 23 batteries represent their lifetime.
}
\source{
Aslam, M. (2021). A new goodness of fit test in the presence of uncertain
parameters. \emph{Complex and Intelligent Systems}, 7(1), 359–365.

Hassan, M. K. and Aslam, M. (2024). Birnbaum Saunders distribution for
imprecise data: statistical properties, estimation methods, and real life
applications, \emph{scientific reports}, 14: 6955.
}
\description{
It is related to the lifetime in 100hours of batteries
}
\examples{
data("batteries")
batteries
}
