% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChildMortalitySA.R
\name{ChildMortalitySA}
\alias{ChildMortalitySA}
\title{Childhood Mortality Rates in Saudi Arabia (1995-2020)}
\format{
A data frame with observations of childhood mortality rates per 1000 live births.
}
\source{
World Health Organization (WHO), Global Health Observatory (GHO) data repository.
}
\description{
Dataset containing childhood mortality rates (under age 5) per 1000 live births for Saudi Arabia from 1995 to 2020. Data sourced from the WHO's global health indicators database.
}
\examples{
data("ChildMortalitySA")
ChildMortalitySA
}
