% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPK.R
\name{CRPK}
\alias{CRPK}
\title{Coronavirus Reproduction Rate Data for Pakistan}
\format{
A data frame with 79 observations of coronavirus reproduction rate in Pakistan from October 2020 to December 2020.
}
\source{
Sherwani, R. A. K., Shakeel, H., Saleem, M., Awan, W. B., Aslam, M.,
Farooq, M. (2021). A new neutrosophic sign test: An application to COVID-19 data,
PLOS ONE, 16(8): e0255671.
}
\description{
Dataset containing the coronavirus reproduction rate in Pakistan from October 2020 to December 2020.
}
\examples{
data("CRPK")
head(CRPK)
}
