% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt.R
\name{debt}
\alias{debt}
\title{Public debt data}
\format{
A data.frame with 23 observations of public debt in Egypt in 200-2022.
}
\source{
Essa, N. I., Zhar, H. M., Abu El-Magd, N. A. T. (2023). Neutrosophic
Generalized Pareto Distribution, \emph{Mathematics and Statistics}, 11(5),
827-833.
}
\description{
It is related to the public debt (in million dollar) in Egypt covering the period from 2000 to 2022.
}
\examples{
data("debt")
debt
}
