% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dioxins.R
\name{dioxins}
\alias{dioxins}
\title{Average Daily Ingestion of Dioxins Data}
\format{
A data frame with 18 observations. Each observation represents the estimated daily ingestion of dioxins (in pg TEQ/day).
}
\source{
Khan, Z., Almazah, M. M. A., Hamood Odhah, O., Alshanbari, H. M. (2022).
Generalized Pareto Model: Properties and Applications in Neutrosophic Data
Modeling, \emph{Mathematical Problems in Engineering}, 2022, 3686968.
}
\description{
The dataset contains estimates of dioxins absorption from the average diet annually for the years 1998 to 2015, as reported in the 2017 annual environmental statistics by the Ministry of Environment, Japan.
}
\examples{
data("dioxins")
dioxins
}
