% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failure.R
\name{failure}
\alias{failure}
\title{Failur time data}
\format{
A data.frame with 30 observations of failure time data on the air conditioning system
}
\source{
Linhart, H. and Zucchini, W. (1986). \emph{Model Selection}, John Wiley & Sons, New York.
}
\description{
It is related to failure times of the cooling system used in an airplane.
}
\examples{
data("failure")
failure
}
