% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICUcovidPK.R
\name{ICUcovidPK}
\alias{ICUcovidPK}
\title{The Daily ICU Occupancy dataset}
\format{
A data frame with observations of daily ICU occupancy:
  - day: Day of December 2020.
  - age_55_and_above: Lower and upper bounds of ICU occupancy for patients aged 55 and above, formatted as "lower-upper".
  - age_between_35_and_55: Lower and upper bounds of ICU occupancy for patients aged between 35 and 55, formatted as "lower-upper".
  - age_below_35: Lower and upper bounds of ICU occupancy for patients aged below 35, formatted as "lower-upper".
  - male: Lower and upper bounds of ICU occupancy for male patients, formatted as "lower-upper".
  - female: Lower and upper bounds of ICU occupancy for female patients, formatted as "lower-upper".
}
\source{
Sherwani, R. A. K., Shakeel, H., Awan, W. B., Faheem, M., Aslam, M. (2021).
Analysis of COVID-19 data using neutrosophic Kruskal Wallis H test, \emph{BMC
Medical Research Methodology}, 21:215.

Sherwani R. A. K., Shakeel, H., Saleem, M., Awan, W. B., Aslam, M., Farooq, M.
(2021). A new neutrosophic sign test: An application to COVID-19 data,
\emph{PLoS ONE}, 16(8): e0255671.
}
\description{
Dataset containing daily ICU occupancy data of COVID-19 patients in Pakistan for December 2020.
}
\examples{
data("ICUcovidPK")
ICUcovidPK
}
