% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRcovidNL.R
\name{MRcovidNL}
\alias{MRcovidNL}
\title{COVID-19 Mortality Rate Data for the Netherlands}
\format{
A data frame with 30 observations of COVID-19 mortality rate for the Netherlands.
}
\source{
Almongy, H. M., Almetwally, E. M., Aljohani, H. M., Alghamdi, A. S. and
Hafez, E. H. (2021). A new extended Rayleigh distribution with applications
of COVID-19 data. \emph{Results in Physics}, 23, 104012.

Jamal, F., Shafiq, S., Aslam, M., Khan, S., Hussain, Z., Abbas, Q. (2024).
Modeling COVID-19 data with a novel neutrosophic Burr-III distribution,
\emph{Scientific Reports}, 14:10810.
}
\description{
Dataset containing COVID-19 mortality rate data recorded in the Netherlands recorded from 31 March to 30 April 2020.
}
\examples{
data("MRcovidNL")
head(MRcovidNL)
}
