% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOx.R
\name{NOx}
\alias{NOx}
\title{Nitrogen Oxides Emissions Data}
\format{
A data frame with 29 observations of nitrogen oxides emissions for Denmark for the period 1990-2018.
}
\source{
Khan, Z., Amin, A., Khan, S. A., Gulistan, M. (2021). Statistical Development
of the Neutrosophic Lognormal Model with Application to Environmental Data,
\emph{Neutrosophic Sets and Systems}, 47, 1-11.
}
\description{
This dataset includes yearly nitrogen oxides emissions for Denmark for the period 1990 to 2018.
}
\examples{
data("NOx")
NOx
}
