% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopVillUSA.R
\name{PopVillUSA}
\alias{PopVillUSA}
\title{Population Villages USA data}
\format{
A data frame with 17 populations of Villages in the USA. Each row represents a village.
}
\source{
Albassam, M., Khan, N., Aslam, M., (2020) The W/S Test for Data Having
Neutrosophic Numbers: An Application to USA Village Population,
\emph{Complexity}, 2020, 3690879.

Rao, G. S., Norouzirad, M., and Mazarei, D. (2023). Neutrosophic
Generalized Exponential Distribution with Application.
\emph{Neutrosophic Sets and Systems}, 55, 471-485.
}
\description{
It is related to population compactness of a few villages in rural USA.
}
\examples{
data("PopVillUSA")
PopVillUSA
}
