% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempLahor.R
\name{tempLahor}
\alias{tempLahor}
\title{Monthly Low and High Temperatures in Lahore, Pakistan (2016–2020)}
\format{
A data frame with 60 rows representing months, and columns for low and high temperatures in Lahore, Pakistan, from 2016 to 2020.
}
\source{
Ahsan-ul-Haq, M., Zafar, J., Aslam, M., Tariq, S. (2024).
Neutrosophic Topp-Leone Distribution for Interval-Valued Data Analysis,
\emph{Journal of Statistical Theory and Applications}, 23, 164–173.
}
\description{
This dataset contains the monthly low and high temperatures recorded in Lahore, Pakistan, over the years 2016 to 2020.
}
\examples{
data("tempLahor")
tempLahor
}
